Move chain agnostic sql types in core/store/models to chainlink-common/sqlutil/models
