+++ core/internal/cltest/factories.go
@@ -2,7 +2,6 @@ package cltest
 
 import (
 	"bytes"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"math/big"
@@ -111,25 +110,6 @@ func WebURL(t testing.TB, unparsed string) models.WebURL {
 	parsed, err := url.Parse(unparsed)
 	require.NoError(t, err)
 	return models.WebURL(*parsed)
-}
-
-// JSONFromString create JSON from given body and arguments
-func JSONFromString(t testing.TB, body string, args ...interface{}) models.JSON {
-	return JSONFromBytes(t, []byte(fmt.Sprintf(body, args...)))
-}
-
-// JSONFromBytes creates JSON from a given byte array
-func JSONFromBytes(t testing.TB, body []byte) models.JSON {
-	j, err := models.ParseJSON(body)
-	require.NoError(t, err)
-	return j
-}
-
-func MustJSONMarshal(t *testing.T, val interface{}) string {
-	t.Helper()
-	bs, err := json.Marshal(val)
-	require.NoError(t, err)
-	return string(bs)
 }
 
 func EmptyCLIContext() *cli.Context {
